python interface: check that NO_TPSA is defined
NO_TPSA (order of TPSA) must be defined because
it defines the type used and changes the template tps.
Therefore it must match the one chosen in the library.

Check that it is defined. Currently python interfance only
used for number of order 1 (linear case). Issue a warning
if a different value is find.